config: pulse-audio: improved the virtual microphone setup
Now the loopback modules, responsible for routing audio from the hardware
microphone and the PRE sink, are disallowed to switch their sources.
This means, even if the user switches the default source,
the loopback modules won’t change their sources.
Additionally, set the virtual microphone as the default input device.